+++ SupportSystemApp.Service/Interface/ISubcategoryService.cs
@@ -11,6 +11,7 @@ namespace SupportSystemApp.Service.Interface
     {
         public List<Subcategory> GetAllByCategoryId(Guid categoryId);
         public Subcategory GetById(Guid id);
+        public Subcategory GetByCategory(Guid categoryId);
         public Subcategory Insert(Subcategory subcategory);
         public Subcategory Update(Subcategory subcategory);
         public Subcategory DeleteById(Guid id);