+++ SupportSystemApp.Web/Controllers/SupportGroupsController.cs
@@ -7,34 +7,31 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using SupportSystemApp.Domain.Domain;
 using SupportSystemApp.Repository;
+using SupportSystemApp.Service.Interface;
 
 namespace SupportSystemApp.Web.Controllers
 {
     public class SupportGroupsController : Controller
     {
-        private readonly ApplicationDbContext _context;
+        private readonly ISupportGroupService _supportGroupService;
 
-        public SupportGroupsController(ApplicationDbContext context)
+        public SupportGroupsController(ISupportGroupService supportGroupService)
         {
-            _context = context;
+            _supportGroupService = supportGroupService;
         }
 
         // GET: SupportGroups
-        public async Task<IActionResult> Index()
+        public IActionResult Index()
         {
-            return View(await _context.SupportGroups.ToListAsync());
+            var supportGroups = _supportGroupService.GetAll();
+            return View(supportGroups);
         }
 
         // GET: SupportGroups/Details/5
-        public async Task<IActionResult> Details(Guid? id)
+        public IActionResult Details(Guid id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            var supportGroup = _supportGroupService.GetById(id);
 
-            var supportGroup = await _context.SupportGroups
-                .FirstOrDefaultAsync(m => m.Id == id);
             if (supportGroup == null)
             {
                 return NotFound();
@@ -54,31 +51,27 @@ namespace SupportSystemApp.Web.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Name,Id")] SupportGroup supportGroup)
+        public IActionResult Create([Bind("Name,Id")] SupportGroup supportGroup)
         {
             if (ModelState.IsValid)
             {
-                supportGroup.Id = Guid.NewGuid();
-                _context.Add(supportGroup);
-                await _context.SaveChangesAsync();
+                _supportGroupService.Insert(supportGroup);
+
                 return RedirectToAction(nameof(Index));
             }
             return View(supportGroup);
         }
 
         // GET: SupportGroups/Edit/5
-        public async Task<IActionResult> Edit(Guid? id)
+        public IActionResult Edit(Guid id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            var supportGroup = _supportGroupService.GetById(id);
 
-            var supportGroup = await _context.SupportGroups.FindAsync(id);
             if (supportGroup == null)
             {
                 return NotFound();
             }
+
             return View(supportGroup);
         }
 
@@ -87,7 +80,7 @@ namespace SupportSystemApp.Web.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(Guid id, [Bind("Name,Id")] SupportGroup supportGroup)
+        public IActionResult Edit(Guid id, [Bind("Name,Id")] SupportGroup supportGroup)
         {
             if (id != supportGroup.Id)
             {
@@ -98,8 +91,7 @@ namespace SupportSystemApp.Web.Controllers
             {
                 try
                 {
-                    _context.Update(supportGroup);
-                    await _context.SaveChangesAsync();
+                    _supportGroupService.Update(supportGroup);
                 }
                 catch (DbUpdateConcurrencyException)
                 {
@@ -118,15 +110,9 @@ namespace SupportSystemApp.Web.Controllers
         }
 
         // GET: SupportGroups/Delete/5
-        public async Task<IActionResult> Delete(Guid? id)
+        public IActionResult Delete(Guid id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var supportGroup = await _context.SupportGroups
-                .FirstOrDefaultAsync(m => m.Id == id);
+            var supportGroup = _supportGroupService.GetById(id);
             if (supportGroup == null)
             {
                 return NotFound();
@@ -138,21 +124,21 @@ namespace SupportSystemApp.Web.Controllers
         // POST: SupportGroups/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(Guid id)
+        public IActionResult DeleteConfirmed(Guid id)
         {
-            var supportGroup = await _context.SupportGroups.FindAsync(id);
+            var supportGroup = _supportGroupService.GetById(id);
+
             if (supportGroup != null)
             {
-                _context.SupportGroups.Remove(supportGroup);
+                _supportGroupService.Delete(id);
             }
 
-            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool SupportGroupExists(Guid id)
         {
-            return _context.SupportGroups.Any(e => e.Id == id);
+            return _supportGroupService.GetById(id) != null;
         }
     }
 }