+++ SupportSystemApp.Web/Controllers/TicketsController.cs
@@ -2,10 +2,12 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using SupportSystemApp.Domain.Domain;
+using SupportSystemApp.Domain.Identity;
 using SupportSystemApp.Repository;
 using SupportSystemApp.Service.Interface;
 
@@ -14,10 +16,17 @@ namespace SupportSystemApp.Web.Controllers
     public class TicketsController : Controller
     {
         private readonly ITicketService _ticketService;
+        private readonly ISiteService _siteService;
+        private readonly ISupportGroupService _supportGroupService;
+        private readonly UserManager<SupportSystemAppUser> _userManager;
 
-        public TicketsController(ITicketService ticketService)
+        public TicketsController(ITicketService ticketService, ISiteService siteService, 
+            ISupportGroupService supportGroupService, UserManager<SupportSystemAppUser> userManager)
         {
             _ticketService = ticketService;
+            _siteService = siteService;
+            _supportGroupService = supportGroupService;
+            _userManager = userManager;
         }
 
         // GET: Tickets
@@ -58,35 +67,104 @@ namespace SupportSystemApp.Web.Controllers
                 return RedirectToAction(nameof(Index));
             }
 
-            //ViewBag.Technitians = new SelectList(_ticketService.GetAll()
-            //    .Select(c => new SelectListItem
-            //    {
-            //        Value = c.Id.ToString(),
-            //        //TO DO
-            //        Text = c.
-            //    })
-            //    .ToList();
+
+            ViewBag.Sites = _siteService.GetAll()
+                .Select(c => new SelectListItem
+                {
+                    Value = c.Id.ToString(),
+                    Text = c.Name
+                })
+                .ToList();
+
+            ViewBag.SupportGroups = _supportGroupService.GetAll()
+                .Select(c => new SelectListItem
+                {
+                    Value = c.Id.ToString(),
+                    Text = c.Name
+                })
+                .ToList();
+
+            ViewBag.Users = _userManager.Users
+                .Select(u => new SelectListItem
+                {
+                    Value = u.Id.ToString(),
+                    Text = u.UserName
+                })
+                .ToList();
+
+            ViewBag.Priorities = Enum.GetValues(typeof(TicketPriority))
+                .Cast<TicketPriority>()
+                .Select(p => new SelectListItem
+                {
+                    Value = p.ToString(),
+                    Text = p.ToString()
+                })
+                .ToList();
+
+            ViewBag.Statuses = Enum.GetValues(typeof(TicketStatus))
+                .Cast<TicketStatus>()
+                .Select(s => new SelectListItem
+                {
+                    Value = s.ToString(),
+                    Text = s.ToString()
+                })
+                .ToList();
 
             return View(ticket);
         }
 
         // GET: Tickets/Edit/5
-        public async Task<IActionResult> Edit(Guid? id)
+        public IActionResult Edit(Guid id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            var ticket = _ticketService.GetById(id);
 
-            var ticket = await _context.Tickets.FindAsync(id);
             if (ticket == null)
             {
                 return NotFound();
             }
-            ViewData["TechnitianId"] = new SelectList(_context.Users, "Id", "Id", ticket.TechnitianId);
-            ViewData["RequesterId"] = new SelectList(_context.Users, "Id", "Id", ticket.RequesterId);
-            ViewData["SiteId"] = new SelectList(_context.Sites, "Id", "Id", ticket.SiteId);
-            ViewData["SupportGroupId"] = new SelectList(_context.SupportGroups, "Id", "Id", ticket.SupportGroupId);
+
+            ViewBag.Sites = _siteService.GetAll()
+                .Select(c => new SelectListItem
+                {
+                    Value = c.Id.ToString(),
+                    Text = c.Name
+                })
+                .ToList();
+
+            ViewBag.SupportGroups = _supportGroupService.GetAll()
+                .Select(c => new SelectListItem
+                {
+                    Value = c.Id.ToString(),
+                    Text = c.Name
+                })
+                .ToList();
+
+            ViewBag.Users = _userManager.Users
+                .Select(u => new SelectListItem
+                {
+                    Value = u.Id.ToString(),
+                    Text = u.UserName
+                })
+                .ToList();
+
+            ViewBag.Priorities = Enum.GetValues(typeof(TicketPriority))
+                .Cast<TicketPriority>()
+                .Select(p => new SelectListItem
+                {
+                    Value = p.ToString(),
+                    Text = p.ToString()
+                })
+                .ToList();
+
+            ViewBag.Statuses = Enum.GetValues(typeof(TicketStatus))
+                .Cast<TicketStatus>()
+                .Select(s => new SelectListItem
+                {
+                    Value = s.ToString(),
+                    Text = s.ToString()
+                })
+                .ToList();
+
             return View(ticket);
         }
 
@@ -95,7 +173,7 @@ namespace SupportSystemApp.Web.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(Guid id, [Bind("TicketNumber,Header,Details,Status,Priority,CreatedAt,DueBy,ResolvedAt,RequesterId,TechnitianId,SiteId,SupportGroupId,Id")] Ticket ticket)
+        public IActionResult Edit(Guid id, [Bind("TicketNumber,Header,Details,Status,Priority,CreatedAt,DueBy,ResolvedAt,RequesterId,TechnitianId,SiteId,SupportGroupId,Id")] Ticket ticket)
         {
             if (id != ticket.Id)
             {
@@ -106,8 +184,7 @@ namespace SupportSystemApp.Web.Controllers
             {
                 try
                 {
-                    _context.Update(ticket);
-                    await _context.SaveChangesAsync();
+                    _ticketService.Update(ticket);
                 }
                 catch (DbUpdateConcurrencyException)
                 {
@@ -122,27 +199,55 @@ namespace SupportSystemApp.Web.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["TechnitianId"] = new SelectList(_context.Users, "Id", "Id", ticket.TechnitianId);
-            ViewData["RequesterId"] = new SelectList(_context.Users, "Id", "Id", ticket.RequesterId);
-            ViewData["SiteId"] = new SelectList(_context.Sites, "Id", "Id", ticket.SiteId);
-            ViewData["SupportGroupId"] = new SelectList(_context.SupportGroups, "Id", "Id", ticket.SupportGroupId);
+            ViewBag.Sites = _siteService.GetAll()
+                .Select(c => new SelectListItem
+                {
+                    Value = c.Id.ToString(),
+                    Text = c.Name
+                })
+                .ToList();
+
+            ViewBag.SupportGroups = _supportGroupService.GetAll()
+                .Select(c => new SelectListItem
+                {
+                    Value = c.Id.ToString(),
+                    Text = c.Name
+                })
+                .ToList();
+
+            ViewBag.Users = _userManager.Users
+                .Select(u => new SelectListItem
+                {
+                    Value = u.Id.ToString(),
+                    Text = u.UserName
+                })
+                .ToList();
+
+            ViewBag.Priorities = Enum.GetValues(typeof(TicketPriority))
+                .Cast<TicketPriority>()
+                .Select(p => new SelectListItem
+                {
+                    Value = p.ToString(),
+                    Text = p.ToString()
+                })
+                .ToList();
+
+            ViewBag.Statuses = Enum.GetValues(typeof(TicketStatus))
+                .Cast<TicketStatus>()
+                .Select(s => new SelectListItem
+                {
+                    Value = s.ToString(),
+                    Text = s.ToString()
+                })
+                .ToList();
             return View(ticket);
         }
 
         // GET: Tickets/Delete/5
-        public async Task<IActionResult> Delete(Guid? id)
+        public IActionResult Delete(Guid id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            var ticket = _ticketService.GetById(id);
 
-            var ticket = await _context.Tickets
-                .Include(t => t.AssignedTo)
-                .Include(t => t.OpenedBy)
-                .Include(t => t.Site)
-                .Include(t => t.SupportGroup)
-                .FirstOrDefaultAsync(m => m.Id == id);
             if (ticket == null)
             {
                 return NotFound();
@@ -154,21 +259,21 @@ namespace SupportSystemApp.Web.Controllers
         // POST: Tickets/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(Guid id)
+        public IActionResult DeleteConfirmed(Guid id)
         {
-            var ticket = await _context.Tickets.FindAsync(id);
+            var ticket = _ticketService.GetById(id);
+
             if (ticket != null)
             {
-                _context.Tickets.Remove(ticket);
+                _ticketService.DeleteById(id);
             }
 
-            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool TicketExists(Guid id)
         {
-            return _context.Tickets.Any(e => e.Id == id);
+            return _ticketService.GetById(id) != null;
         }
     }
 }